[CLI] Move util functions and move spec / model functions to separate file.
