+++ cli/src/pcluster/utils.py
@@ -123,6 +123,18 @@ def camelcase(snake_case_word):
     return "".join(word.title() for word in parts)
 
 
+def to_kebab_case(input):
+    """Convert a string into its snake case representation."""
+    str1 = re.sub("(.)([A-Z][a-z]+)", r"\1-\2", input).replace('_', '-')
+    return re.sub("([a-z0-9])([A-Z])", r"\1-\2", str1).lower()
+
+
+def to_snake_case(input):
+    """Convert a string into its snake case representation."""
+    str1 = re.sub("(.)([A-Z][a-z]+)", r"\1_\2", input).replace('-', '_')
+    return re.sub("([a-z0-9])([A-Z])", r"\1_\2", str1).lower()
+
+
 def get_stack_output_value(stack_outputs, output_key):
     """
     Get output value from Cloudformation Stack Output.
